# docs/swagger/swagger.yaml
info:
  contact: {}
  license: {}
paths:
...
  /api/v1/problem/delete:
    post:
      parameters:
      - description: id_list
        in: query
        name: id_list
        required: true
        type: json
      produces:
      - application/json
      summary: Delete Problem
  /api/v1/problem/detail:
    get:
      parameters:
      - description: problem_id
        in: query
        name: problem_id
        required: true
        type: integer
      produces:
      - application/json
      summary: Problem Detail
  /api/v1/problem/edit:
    post:
      parameters:
      - description: title
        in: query
        name: title
        required: true
        type: string
      - description: author
        in: query
        name: author
        required: true
        type: string
      - description: description
        in: query
        name: description
        required: true
        type: string
      - description: input
        in: query
        name: input
        required: true
        type: string
      - description: output
        in: query
        name: output
        required: true
        type: string
      - description: sample_input
        in: query
        name: sample_input
        required: true
        type: string
      - description: sample_output
        in: query
        name: sample_output
        required: true
        type: string
      - description: hint
        in: query
        name: hint
        type: string
      produces:
      - application/json
      summary: Update Problem
...
  /api/v1/problem/new:
    post:
      parameters:
      - description: title
        in: query
        name: title
        required: true
        type: string
      - description: author
        in: query
        name: author
        required: true
        type: string
      - description: description
        in: query
        name: description
        required: true
        type: string
      - description: input
        in: query
        name: input
        required: true
        type: string
      - description: output
        in: query
        name: output
        required: true
        type: string
      - description: sample_input
        in: query
        name: sample_input
        required: true
        type: string
      - description: sample_output
        in: query
        name: sample_output
        required: true
        type: string
      - description: hint
        in: query
        name: hint
        type: string
      produces:
      - application/json
      summary: Create Problem
  /api/v1/problem/submit:
    post:
      parameters:
      - description: pid
        in: query
        name: pid
        required: true
        type: json
      - description: uid
        in: query
        name: uid
        required: true
        type: json
      - description: code
        in: query
        name: code
        required: true
        type: json
      - description: memory
        in: query
        name: memory
        required: true
        type: json
      - description: language
        in: query
        name: language
        required: true
        type: json
      produces:
      - application/json
      summary: Submit Problem
...
swagger: "2.0"